separated test file into two
